application_source/python-otel-example/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.128.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
httpx>=0.25.0
httptools==0.7.1
idna==3.11
importlib_metadata==8.7.1
opentelemetry-api==1.39.1
opentelemetry-distro==0.60b1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-asyncio==0.60b1
opentelemetry-instrumentation-click==0.60b1
opentelemetry-instrumentation-dbapi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-instrumentation-grpc==0.60b1
opentelemetry-instrumentation-httpx>=0.41b0
opentelemetry-instrumentation-logging==0.60b1
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-instrumentation-sqlite3==0.60b1
opentelemetry-instrumentation-starlette==0.60b1
opentelemetry-instrumentation-threading==0.60b1
opentelemetry-instrumentation-tortoiseorm==0.60b1
opentelemetry-instrumentation-urllib==0.60b1
opentelemetry-instrumentation-urllib3==0.60b1
opentelemetry-instrumentation-wsgi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
packaging==25.0
protobuf==6.33.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
zipp==3.23.0
